Prevent unicode characters breaking wda/source call
Unicode characters previously were breaking wdaproxy. This fix prevents that from happening. Weirdly what was crashing it was outputting those characters into the log. Without that it works fine. It seems that the log output library simply explodes when handed those characters. The fix is to attempt to convert to utf8, discarding multi-byte characters... weirdly this seems to convert all the characters. Even if it doesn't it doesn't really matter as it is only removing them from the log. Better to remove the characters from the log than to crash.